+++ tests/html/flow/HrTest.php
@@ -1,0 +1,154 @@
+<?php
+
+declare(strict_types=1);
+
+namespace yii\ui\tests\html\flow;
+
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\TestCase;
+use yii\ui\html\flow\Hr;
+use yii\ui\tests\support\stub\DefaultProvider;
+use yii\ui\tests\support\TestSupport;
+use yii\ui\values\Direction;
+
+#[Group('html')]
+#[Group('flow')]
+#[Group('Hr')]
+final class HrTest extends TestCase
+{
+    use TestSupport;
+
+    public function testRenderWithAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr class="test-class">
+            HTML,
+            Hr::tag()->attributes(['class' => 'test-class'])->render(),
+            "Failed asserting that element renders correctly with 'attributes()' method.",
+        );
+    }
+
+    public function testRenderWithClass(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr class="test-class">
+            HTML,
+            Hr::tag()->class('test-class')->render(),
+            "Failed asserting that element renders correctly with 'class' attribute.",
+        );
+    }
+
+    public function testRenderWithDataAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr data-value="test-value">
+            HTML,
+            Hr::tag()->dataAttributes(['value' => 'test-value'])->render(),
+            "Failed asserting that element renders correctly with 'dataAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithDefaultConfigurationValues(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr class="default-class" title="default-title">
+            HTML,
+            Hr::tag(['class' => 'default-class', 'title' => 'default-title'])->render(),
+            'Failed asserting that default configuration values are applied correctly.',
+        );
+    }
+
+    public function testRenderWithDefaultProvider(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr class="default-provider">
+            HTML,
+            Hr::tag()->addDefaultProvider(DefaultProvider::class)->render(),
+            'Failed asserting that default provider is applied correctly.',
+        );
+    }
+
+    public function testRenderWithDefaultValues(): void
+    {
+        $instance = Hr::tag();
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr>
+            HTML,
+            $instance->render(),
+            'Failed asserting that element renders correctly with default values.',
+        );
+    }
+
+    public function testRenderWithDir(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr dir="rtl">
+            HTML,
+            Hr::tag()->dir(Direction::RTL)->render(),
+            "Failed asserting that element renders correctly with 'dir' attribute.",
+        );
+    }
+
+    public function testRenderWithHidden(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr hidden>
+            HTML,
+            Hr::tag()->hidden(true)->render(),
+            "Failed asserting that element renders correctly with 'hidden' attribute.",
+        );
+    }
+
+    public function testRenderWithId(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr id="test-id">
+            HTML,
+            Hr::tag()->id('test-id')->render(),
+            "Failed asserting that element renders correctly with 'id' attribute.",
+        );
+    }
+
+    public function testRenderWithLang(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr lang="es">
+            HTML,
+            Hr::tag()->lang('es')->render(),
+            "Failed asserting that element renders correctly with 'lang' attribute.",
+        );
+    }
+
+    public function testRenderWithStyle(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr style="test-value">
+            HTML,
+            Hr::tag()->style('test-value')->render(),
+            "Failed asserting that element renders correctly with 'style' attribute.",
+        );
+    }
+
+    public function testRenderWithTitle(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <hr title="test-value">
+            HTML,
+            Hr::tag()->title('test-value')->render(),
+            "Failed asserting that element renders correctly with 'title' attribute.",
+        );
+    }
+}